+++ Model/Transaction/Refund.php
@@ -59,6 +59,7 @@ class Refund extends \Taxjar\SalesTax\Model\Transaction
 
         $refund = [
             'plugin' => 'magento',
+            'provider' => $this->getProvider($order),
             'transaction_id' => $creditmemo->getIncrementId() . '-refund',
             'transaction_reference_id' => $order->getIncrementId(),
             'transaction_date' => $creditmemo->getCreatedAt(),
@@ -76,8 +77,26 @@ class Refund extends \Taxjar\SalesTax\Model\Transaction
         );
 
         if (isset($this->request['line_items'])) {
-            foreach ($this->request['line_items'] as $lineItem) {
+            $adjustmentFee = $creditmemo->getAdjustmentNegative();
+            $adjustmentRefund = $creditmemo->getAdjustmentPositive();
+
+            // Discounts on credit memos act as fees and shouldn't be included in $itemDiscounts
+            foreach ($this->request['line_items'] as $k => $lineItem) {
+                $lineItemSubtotal = $lineItem['unit_price'] * $lineItem['quantity'];
+                $this->request['line_items'][$k]['discount'] += ($adjustmentFee * ($lineItemSubtotal / $subtotal));
                 $itemDiscounts += $lineItem['discount'];
+            }
+
+            if ($adjustmentRefund) {
+                $this->request['line_items'][] = [
+                    'id' => 'adjustment-refund',
+                    'quantity' => 1,
+                    'product_identifier' => 'adjustment-refund',
+                    'description' => 'Adjustment Refund',
+                    'unit_price' => $adjustmentRefund,
+                    'discount' => 0,
+                    'sales_tax' => 0
+                ];
             }
         }