+++ Setup/UpgradeData.php
@@ -20,24 +20,75 @@ namespace Taxjar\SalesTax\Setup;
 use Magento\Customer\Api\CustomerMetadataInterface;
 use Magento\Eav\Model\Config;
 use Magento\Eav\Setup\EavSetup;
-use Magento\Framework\Setup\UpgradeDataInterface;
-use Magento\Framework\Setup\ModuleDataSetupInterface;
 use Magento\Framework\Setup\ModuleContextInterface;
+use Magento\Framework\Setup\ModuleDataSetupInterface;
+use Magento\Framework\App\Config\ScopeConfigInterface;
+use Magento\Framework\Setup\UpgradeDataInterface;
+use Taxjar\SalesTax\Model\ClientFactory;
+use Taxjar\SalesTax\Model\Configuration as TaxjarConfig;
 
 class UpgradeData implements UpgradeDataInterface
 {
-    private $eavSetupFactory;
+    /**
+     * @var \Magento\Eav\Model\AttributeRepository
+     */
     private $attributeRepository;
+
+    /**
+     * @var \Taxjar\SalesTax\Model\Client
+     */
+    private $client;
+
+    /**
+     * @var Config
+     */
     private $eavConfig;
 
+    /**
+     * @var \Magento\Eav\Setup\EavSetupFactory
+     */
+    private $eavSetupFactory;
+
+    /**
+     * @var \Magento\Framework\Event\ManagerInterface
+     */
+    private $eventManager;
+
+    /**
+     * @var \Magento\Config\Model\ResourceModel\Config
+     */
+    private $resourceConfig;
+
+    /**
+     * @var \Magento\Framework\App\Config\ScopeConfigInterface
+     */
+    private $scopeConfig;
+
+    /**
+     * @param \Magento\Eav\Model\AttributeRepository $attributeRepository
+     * @param ClientFactory $clientFactory
+     * @param Config $eavConfig
+     * @param \Magento\Eav\Setup\EavSetupFactory $eavSetupFactory
+     * @param \Magento\Framework\Event\ManagerInterface $eventManager
+     * @param \Magento\Config\Model\ResourceModel\Config $resourceConfig
+     * @param ScopeConfigInterface $scopeConfig
+     */
     public function __construct(
-        \Magento\Eav\Setup\EavSetupFactory $eavSetupFactory,
         \Magento\Eav\Model\AttributeRepository $attributeRepository,
-        Config $eavConfig
+        ClientFactory $clientFactory,
+        Config $eavConfig,
+        \Magento\Eav\Setup\EavSetupFactory $eavSetupFactory,
+        \Magento\Framework\Event\ManagerInterface $eventManager,
+        \Magento\Config\Model\ResourceModel\Config $resourceConfig,
+        ScopeConfigInterface $scopeConfig
     ) {
-        $this->eavSetupFactory = $eavSetupFactory;
         $this->attributeRepository = $attributeRepository;
+        $this->client = $clientFactory->create();
         $this->eavConfig = $eavConfig;
+        $this->eavSetupFactory = $eavSetupFactory;
+        $this->eventManager = $eventManager;
+        $this->resourceConfig = $resourceConfig;
+        $this->scopeConfig = $scopeConfig;
     }
 
     public function upgrade(
@@ -162,6 +213,26 @@ class UpgradeData implements UpgradeDataInterface
                 $lastSyncCode);
             $lastSyncCodeId->setData('used_in_forms', ['adminhtml_customer',]);
             $lastSyncCodeId->getResource()->save($lastSyncCodeId);
+        }
+
+        if (version_compare($context->getVersion(), '1.0.3', '<')) {
+            if ($this->scopeConfig->getValue(TaxjarConfig::TAXJAR_APIKEY)) {
+                $configJson = $this->client->getResource('config');
+
+                if (is_array($configJson) && isset($configJson['configuration']) && isset($configJson['configuration']['states'])) {
+                    $this->resourceConfig->saveConfig(
+                        TaxjarConfig::TAXJAR_STATES,
+                        json_encode(explode(',', $configJson['configuration']['states'])),
+                        'default',
+                        0
+                    );
+                }
+            }
+        }
+
+        if (version_compare($context->getVersion(), '1.0.4', '<')) {
+            $this->resourceConfig->deleteConfig('tax/taxjar/categories', 'default', '');
+            $this->eventManager->dispatch('taxjar_salestax_import_categories');
         }
     }
 }